Refactor SDK types and share Card interface with UI
Fixes #649
Fixes #631

Change-type: patch
