+++ lib/sdk/index.ts
@@ -18,18 +18,65 @@ import axios, { AxiosRequestConfig, AxiosResponse, CancelTokenSource } from 'axi
 import * as Bluebird from 'bluebird';
 import { JSONSchema6 } from 'json-schema';
 import * as _ from 'lodash';
+import { Card } from '../Types';
 import { AuthSdk } from './auth';
 import { CardSdk } from './card';
 import { JellyfishStream, JellyfishStreamManager } from './stream';
-import {
-	ActionResponse,
-	Card,
-	QueryOptions,
-	SDKInterface,
-	SDKOptions,
-	ServerResponse
-} from './Types';
 import * as utils from './utils';
+
+export { JellyfishStream };
+
+export type ActionResponse<D = any> = {
+	error: false;
+	data: D;
+} | {
+	error: true;
+	data: {
+		message: string;
+	};
+};
+
+export interface QueryOptions {
+	limit?: number;
+	skip?: number;
+	sortBy?: string | string[];
+	sortDir?: 'asc' | 'desc';
+}
+
+export interface ServerResponse {
+	error: boolean;
+	data: any;
+}
+
+export interface SDKOptions {
+	apiUrl: string;
+	apiPrefix: string;
+	authToken?: string;
+}
+
+export interface SDKInterface {
+	getApiUrl: () => string | undefined;
+	getAuthToken: () => string | undefined;
+	setAuthToken: (token: string) => void;
+	clearAuthToken: () => void;
+
+	cancelAllRequests: (reason?: string) => void;
+	cancelAllStreams: () => void;
+
+	action: <D = any>(body: {
+		card: string;
+		action: string;
+		arguments?: any;
+	}) => Bluebird<D>;
+
+	query: <T extends Card>(schema: JSONSchema6) => Bluebird<T[]>;
+
+	post: <R = ServerResponse>(endpoint: string, body: any, options?: AxiosRequestConfig) => Bluebird<AxiosResponse<R> | void>;
+
+	card: {
+		get: (idOrSlug: string, options?: object) => Bluebird<Card | null>;
+	};
+}
 
 const trimSlash = (s: string) => _.trim(s, '/');
