+++ lib/ui/components/TailStreamer.tsx
@@ -1,7 +1,7 @@
 import { JSONSchema6 } from 'json-schema';
 import * as _ from 'lodash';
 import * as React from 'react';
-import { JellyfishStream } from '../../sdk/stream';
+import { JellyfishStream } from '../../sdk';
 import { Card } from '../../Types';
 import { store } from '../core';
 import { sdk } from '../core/sdk';
